kiff.py: do not use layer-names to guess .pdf file-names
in kicad_cli we have use the _untranslated_ layer name (F.SilkS)
to select the layers to plot. However it uses the user-defined
layer name (F.Silkscreen) in the file-name of the .pdf it generates.

Instead of dealing with this inconsitency, I just use a *.pdf glob
in the first output folder to get the names of the .pdf files to load.